new attr `purge_meta_keys`, correctely handle comments
The new `purge_meta_keys` attribue allows to fully manage metadata
keys, which means, that unspecified metadata keys are now removed,
this attribue is set to true.

Comments are now managed in a much better way, which allows adding
and removing comments.

Still an open issue: when and which comment character should be
used? Currently hard coded to '#'